Remove the DependencyFileParser worker
Instead, do that work in the DependencyFileFetcher worker.
